refactor: 인증 관련 유틸 함수 분리
- getAuthToken: 인증 토큰 반환 함수
- redirectIfNotLoggedIn: 로그인페이지 리디렉션 함수
