+++ src/api/dashboardData.js
@@ -1,19 +1,9 @@
+import { getAuthToken, redirectIfNotLoggedIn } from '../utils/auth';
+
 /** 대시보드 API 요청 */
 export async function fetchDashboardData() {
-  // TODO: 로그인 구현되면 아래 구문으로 교체
-  // const token = localStorage.getItem('token');
-  const token = '임시 토큰';
-
-  if (!token) {
-    const answer = confirm('로그인이 필요합니다. 로그인 페이지로 이동할까요?');
-
-    const redirectUrl = answer
-      ? '/src/pages/LoginAndSignUp/LoginAndSignUp.html'
-      : '/src/pages/Community/Community.html';
-
-    window.location.replace(redirectUrl);
-    return;
-  }
+  const token = getAuthToken();
+  if (!token) redirectIfNotLoggedIn();
 
   const res = await fetch('https://server.bookmark.soop.run/dashboard', {
     method: 'GET',
