+++ src/utils/auth.js
@@ -1,0 +1,16 @@
+/** 인증 토큰 반환 함수
+ * - 토큰이 없을 시(비로그인) 빈 문자열 반환
+ */
+export function getAuthToken() {
+  return localStorage.getItem('token') ?? '';
+}
+
+/** 비로그인 상태 시 로그인 페이지로 리디렉션하는 함수 */
+export function redirectIfNotLoggedIn() {
+  const answer = confirm('로그인이 필요합니다. 로그인 페이지로 이동할까요?');
+  const redirectUrl = answer
+    ? '/src/pages/LoginAndSignUp/LoginAndSignUp.html'
+    : '/src/pages/Community/Community.html';
+
+  window.location.replace(redirectUrl);
+}
